+++ CHANGELOG.md
@@ -66,6 +66,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added global options to define the PCB details (`pcb_material`,
   `solder_mask_color`, `silk_screen_color` and `pcb_finish`)
 - Report generation (for design house) (#93)
+- New output to print PCB layers in SVG format.
 
 ### Changed
 - Internal BoM: now components with different Tolerance, Voltage, Current